src/main/java/com/eguaks/core/adapters/LocalDateTypeAdapter.java
package com.eguaks.core.adapters;

import javax.xml.bind.annotation.adapters.XmlAdapter;
import java.time.LocalDate;
import java.time.format.DateTimeFormatter;

/**
 * Created by JohnSigvald on 10/05/2014.
 */
public class LocalDateTypeAdapter extends XmlAdapter<String, LocalDate> {

    @Override
    public LocalDate unmarshal(String v) throws Exception {
        return LocalDate.parse(v);
    }

    @Override
    public String marshal(LocalDate v) throws Exception {
        DateTimeFormatter formatter = DateTimeFormatter.ISO_LOCAL_DATE;
        return v.format(formatter);
    }
}
